+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Version 155:
+
+* Fix memory leak in advanced server examples
+
+--------------------------------------------------------------------------------
+
 Version 154:
 
 * Type check completion handlers